Rename `render-dom` and `render-ssr` to `render_dom` and `render_ssr`
