+++ src/compile/nodes/shared/AbstractBlock.ts
@@ -1,4 +1,4 @@
-import Block from '../../render-dom/Block';
+import Block from '../../render_dom/Block';
 import Component from './../../Component';
 import Node from './Node';
 import { INode } from '../interfaces';
