+++ src/compile/render_dom/wrappers/InlineComponent/index.ts
