+++ src/compile/render_dom/wrappers/Window.ts
