+++ src/compile/render_ssr/handlers/InlineComponent.ts
