+++ src/compile/render_ssr/handlers/shared/get_slot_scope.ts
